server/pyproject.toml
[project]
name = "bandcorder"
version = "2020.0.0"
dependencies = [
  "sounddevice==0.5.1",
  "soundfile==0.12.1",
  "numpy==2.1.3",
  "tornado==6.4.2",
  "python-socketio==5.11.4",
  "pyyaml==6.0.2",
]
requires-python = ">=3.8"

[tool.setuptools]
py-modules = []